+++ Workflow/Sources/Workflow/Models/Workflow.swift
@@ -21,11 +21,196 @@ import Foundation
 ///         .thenProceed(with: SomeOtherFlowRepresentable.self)
 /// }
 /// ```
-public final class Workflow<F: FlowRepresentable>: AnyWorkflow {
-    public required init(_ node: AnyWorkflow.Element?) {
+public final class Workflow<F: FlowRepresentable>: LinkedList<_WorkflowItem> {
+    public required init(_ node: Element?) {
         super.init(node)
     }
 
+    /// The `OrchestartionResponder` the `Workflow` will send actions to.
+    public internal(set) var orchestrationResponder: OrchestrationResponder?
+
+    /// Creates an `Workflow` with a `WorkflowItem` that has metadata, but no instance
+    public convenience init(_ metadata: FlowRepresentableMetadata) {
+        self.init(Node(with: _WorkflowItem(metadata: metadata, instance: nil)))
+    }
+
+    /// Appends a `WorkflowItem` that has metadata, but no instance
+    public func append(_ metadata: FlowRepresentableMetadata) {
+        append(_WorkflowItem(metadata: metadata, instance: nil))
+    }
+
+    /**
+     Launches the `Workflow`.
+     - Parameter orchestrationResponder: The `OrchestrationResponder` to notify when the `Workflow` proceeds or backs up.
+     - Parameter launchStyle: The launch style to use.
+     - Parameter onFinish: The closure to call when the last element in the workflow proceeds; called with the `AnyWorkflow.PassedArgs` the workflow finished with.
+     - Returns: The first loaded instance or nil, if none was loaded.
+     */
+    @discardableResult public func launch(withOrchestrationResponder orchestrationResponder: OrchestrationResponder,
+                                          launchStyle: LaunchStyle = .default,
+                                          onFinish: ((AnyWorkflow.PassedArgs) -> Void)? = nil) -> Element? {
+        launch(withOrchestrationResponder: orchestrationResponder,
+               passedArgs: .none,
+               launchStyle: launchStyle,
+               onFinish: onFinish)
+    }
+
+    /**
+     Launches the `Workflow`.
+
+     ### Discussion
+     Args are passed to the first instance, it has the opportunity to load, not load and transform them, or just not load.
+     In the event an instance does not load and does not transform args, they are passed unmodified to the next instance in the `Workflow` until one loads.
+
+     - Parameter orchestrationResponder: The `OrchestrationResponder` to notify when the `Workflow` proceeds or backs up.
+     - Parameter args: The arguments to pass to the first instance(s).
+     - Parameter launchStyle: The launch style to use.
+     - Parameter onFinish: The closure to call when the last element in the workflow proceeds; called with the `AnyWorkflow.PassedArgs` the workflow finished with.
+     - Returns: The first loaded instance or nil, if none was loaded.
+     */
+    @discardableResult public func launch(withOrchestrationResponder orchestrationResponder: OrchestrationResponder,
+                                          args: Any?,
+                                          withLaunchStyle launchStyle: LaunchStyle = .default,
+                                          onFinish: ((AnyWorkflow.PassedArgs) -> Void)? = nil) -> Element? {
+        launch(withOrchestrationResponder: orchestrationResponder,
+               passedArgs: .args(args),
+               launchStyle: launchStyle,
+               onFinish: onFinish)
+    }
+
+    /**
+     Launches the `Workflow`.
+
+     ### Discussion
+     passedArgs are passed to the first instance, it has the opportunity to load, not load and transform them, or just not load.
+     In the event an instance does not load and does not transform args, they are passed unmodified to the next instance in the `Workflow` until one loads.
+
+     - Parameter orchestrationResponder: The `OrchestrationResponder` to notify when the `Workflow` proceeds or backs up.
+     - Parameter passedArgs: The arguments to pass to the first instance(s).
+     - Parameter launchStyle: The launch style to use.
+     - Parameter onFinish: The closure to call when the last element in the workflow proceeds; called with the `AnyWorkflow.PassedArgs` the workflow finished with.
+     - Returns: The first loaded instance or nil, if none was loaded.
+     */
+    @discardableResult public func launch(withOrchestrationResponder orchestrationResponder: OrchestrationResponder,
+                                          passedArgs: AnyWorkflow.PassedArgs,
+                                          launchStyle: LaunchStyle = .default,
+                                          onFinish: ((AnyWorkflow.PassedArgs) -> Void)? = nil) -> Element? {
+        removeInstances()
+        self.orchestrationResponder = orchestrationResponder
+        var root: Element?
+        var passedArgs = passedArgs
+
+        let firstLoadedInstance = first?.traverse { [self] nextNode in
+            let flowRepresentable = nextNode.value.metadata.flowRepresentableFactory(passedArgs)
+            flowRepresentable.workflow = AnyWorkflow(self)
+            flowRepresentable.proceedInWorkflowStorage = { passedArgs = $0 }
+
+            let shouldLoad = flowRepresentable.shouldLoad()
+
+            defer {
+                let persistence = nextNode.value.metadata.calculatePersistence(passedArgs)
+                if shouldLoad {
+                    nextNode.value.instance = flowRepresentable
+                    setupCallbacks(for: nextNode, onFinish: onFinish)
+                } else if !shouldLoad && persistence == .persistWhenSkipped {
+                    nextNode.value.instance = flowRepresentable
+                    setupCallbacks(for: nextNode, onFinish: onFinish)
+                    orchestrationResponder.launchOrProceed(to: nextNode, from: root)
+                    root = nextNode
+                }
+            }
+            return shouldLoad
+        }
+
+        guard let first = firstLoadedInstance  else {
+            onFinish?(passedArgs)
+            return nil
+        }
+
+        orchestrationResponder.launchOrProceed(to: first, from: root)
+
+        return firstLoadedInstance
+    }
+
+    public func _abandon() {
+        removeInstances()
+        orchestrationResponder = nil
+    }
+
+    deinit {
+        removeInstances()
+        orchestrationResponder = nil
+    }
+
+    private func removeInstances() {
+        forEach { node in
+            node.value.instance?.proceedInWorkflowStorage = nil
+            node.value.instance = nil
+        }
+    }
+
+    private func setupProceedCallbacks(_ node: Element, _ onFinish: ((AnyWorkflow.PassedArgs) -> Void)?) {
+        node.value.instance?.proceedInWorkflowStorage = { [weak self] args in
+            guard let self = self else { return }
+            self.setupProceedInWorkflowStorage(node: node,
+                                               args: args,
+                                               onFinish: onFinish)
+        }
+    }
+
+    private func setupProceedInWorkflowStorage(node: Element, args: AnyWorkflow.PassedArgs, onFinish: ((AnyWorkflow.PassedArgs) -> Void)?) {
+        var argsToPass = args
+        var root: Element?
+        // traverse AND mutate the above variables
+        let nextLoadedNode = node.next?.traverse { nextNode in
+            let persistence = nextNode.value.metadata.calculatePersistence(argsToPass)
+            let flowRepresentable = nextNode.value.metadata.flowRepresentableFactory(argsToPass)
+            flowRepresentable.workflow = AnyWorkflow(self)
+
+            // Capture new arguments, if shouldLoad calls proceedInWorkflow
+            flowRepresentable.proceedInWorkflowStorage = { argsToPass = $0 }
+
+            let shouldLoad = flowRepresentable.shouldLoad()
+            nextNode.value.instance = (shouldLoad || (!shouldLoad && persistence == .persistWhenSkipped)) ? flowRepresentable : nil
+
+            if !shouldLoad && persistence == .persistWhenSkipped {
+                nextNode.value.instance = flowRepresentable
+                setupCallbacks(for: nextNode, onFinish: onFinish)
+                orchestrationResponder?.proceed(to: nextNode, from: root ?? node)
+                root = nextNode
+            }
+
+            return shouldLoad
+        }
+
+        guard let nextNode = nextLoadedNode else {
+            onFinish?(argsToPass)
+            return
+        }
+
+        setupCallbacks(for: nextNode, onFinish: onFinish)
+        orchestrationResponder?.proceed(to: nextNode, from: root ?? node)
+    }
+
+    private func setupBackUpCallbacks(_ node: Element, _ onFinish: ((AnyWorkflow.PassedArgs) -> Void)?) {
+        node.value.instance?.backUpInWorkflowStorage = { [self] in
+            let previousLoadedNode = node.traverse(direction: .backward) { previousNode in
+                previousNode.value.instance != nil
+            }
+
+            guard let previousNode = previousLoadedNode else { throw WorkflowError.failedToBackUp }
+
+            orchestrationResponder?.backUp(from: node, to: previousNode)
+        }
+    }
+
+    private func setupCallbacks(for node: Element, onFinish: ((AnyWorkflow.PassedArgs) -> Void)?) {
+        setupProceedCallbacks(node, onFinish)
+        setupBackUpCallbacks(node, onFinish)
+    }
+}
+
+extension Workflow {
     /**
      A way of creating workflows with a fluent API. Useful for complex workflows with difficult requirements
      - Parameter type: A reference to the class used to create the workflow
