+++ Workflow/Sources/WorkflowUIKit/Extensions/WorkflowUIKitAdditions.swift
@@ -73,6 +73,18 @@ extension Workflow {
     }
 }
 
+extension Workflow {
+    /**
+    Called when the workflow should be terminated, and the app should return to the point before the workflow was launched
+    - Parameter animated: A boolean indicating whether abandoning the workflow should be animated
+    - Parameter onFinish: A callback after the workflow has been abandoned.
+    - Note: In order to dismiss UIKit views the workflow must have an `OrchestrationResponder` that is a `UIKitPresenter`.
+    */
+    public func abandon(animated: Bool = true, onFinish:(() -> Void)? = nil) {
+        AnyWorkflow(self).abandon(animated: animated, onFinish: onFinish)
+    }
+}
+
 extension AnyWorkflow {
     /**
     Called when the workflow should be terminated, and the app should return to the point before the workflow was launched
